Fix node refresh so that it keeps expansion state
VS Code released an [update today](https://code.visualstudio.com/updates/v1_20#_custom-views) which adds an 'id' property to tree items. This id tells VS Code that a node is the same even if the label changes.

I also made the id optional - we've had a bit of debate about whether or not to have id on our tree items and I think the best solution is to follow VS Code's pattern.

Bumped the minor version since this requires an update to the VS Code engine version.