Allow undefined in telemetry properties (#293)
It's bothered me for a while that telemetry properties were required. For example some times in testing it would be easier to mock IActionContext if these were optional. Also some times we set telemetry properties to a value that might be undefined. It would be easier if we didn't have to cast that to a string.