Go all-in on `IActionContext` (#503)
Now wizards, the tree item picker, and any command/event will leverage actionContext as their base "context" object (which can obviously be extended as desired).

Since I was using `actionContext` so much, I decided to make it the first parameter for any command, rather than using `this`. The only reason I originally used `this` was because people could essentially ignore the `this` parameter if they didn't want to use actionContext. However, that was always confusing and led to things like `const me = this;`.