Allow a repository to specify additional setup steps
Right now any repository consuming these templates can customize how
their repository is build, tested, and published by changing the
behavior of 'npm run build', test, etc. But if you need some additional
setup tasks that don't cleanly fit into npm then it's tough to do.

The specific motivation for this change is wanting to use
Nerdbank.GitVersioning to update our Azure DevOps build numbers to match
what it's computing in an internal repository consuming these templates.
The support for that though requires installing a .NET based tool onto
the machine, which means we need to ensure .NET is installed on the
machine. Rather than reinventing that, it'd be best to use the existing
Azure DevOps task which already handles that.

The alternative approaches here would be to:

1. Just have the templates always install .NET, but that's just a waste
   of time for other repositories.
2. Add a specific setting to enable installing .NET, but rather than
   creating a one-off feature, it's just as easy to make a general
   mechanism.
3. Stop having the other repository consume these templates at all.